Refactor search.search pkg to search.
