+++ benchmarks/src/main/java/io/druid/benchmark/query/SearchBenchmark.java
@@ -63,9 +63,9 @@ import io.druid.query.search.SearchQueryQueryToolChest;
 import io.druid.query.search.SearchQueryRunnerFactory;
 import io.druid.query.search.SearchResultValue;
 import io.druid.query.search.SearchStrategySelector;
-import io.druid.query.search.search.SearchHit;
-import io.druid.query.search.search.SearchQuery;
-import io.druid.query.search.search.SearchQueryConfig;
+import io.druid.query.search.SearchHit;
+import io.druid.query.search.SearchQuery;
+import io.druid.query.search.SearchQueryConfig;
 import io.druid.query.spec.MultipleIntervalSegmentSpec;
 import io.druid.query.spec.QuerySegmentSpec;
 import io.druid.segment.IncrementalIndexSegment;
