+++ processing/src/main/java/io/druid/query/DefaultGenericQueryMetricsFactory.java
@@ -25,6 +25,7 @@ import com.google.inject.Inject;
 import io.druid.guice.LazySingleton;
 import io.druid.guice.annotations.Json;
 import io.druid.jackson.DefaultObjectMapper;
+import io.druid.query.search.SearchQueryConfig;
 
 @LazySingleton
 public class DefaultGenericQueryMetricsFactory implements GenericQueryMetricsFactory
@@ -35,7 +36,7 @@ public class DefaultGenericQueryMetricsFactory implements GenericQueryMetricsFac
   /**
    * Should be used only in tests, directly or indirectly (e. g. in {@link
    * io.druid.query.search.SearchQueryQueryToolChest#SearchQueryQueryToolChest(
-   * io.druid.query.search.search.SearchQueryConfig, IntervalChunkingQueryRunnerDecorator)}).
+   * SearchQueryConfig, IntervalChunkingQueryRunnerDecorator)}).
    */
   @VisibleForTesting
   public static GenericQueryMetricsFactory instance()
