+++ processing/src/main/java/io/druid/query/Druids.java
@@ -42,12 +42,12 @@ import io.druid.query.filter.SelectorDimFilter;
 import io.druid.query.metadata.metadata.ColumnIncluderator;
 import io.druid.query.metadata.metadata.SegmentMetadataQuery;
 import io.druid.query.search.SearchResultValue;
-import io.druid.query.search.search.ContainsSearchQuerySpec;
-import io.druid.query.search.search.FragmentSearchQuerySpec;
-import io.druid.query.search.search.InsensitiveContainsSearchQuerySpec;
-import io.druid.query.search.search.SearchQuery;
-import io.druid.query.search.search.SearchQuerySpec;
-import io.druid.query.search.search.SearchSortSpec;
+import io.druid.query.search.ContainsSearchQuerySpec;
+import io.druid.query.search.FragmentSearchQuerySpec;
+import io.druid.query.search.InsensitiveContainsSearchQuerySpec;
+import io.druid.query.search.SearchQuery;
+import io.druid.query.search.SearchQuerySpec;
+import io.druid.query.search.SearchSortSpec;
 import io.druid.query.select.PagingSpec;
 import io.druid.query.select.SelectQuery;
 import io.druid.query.spec.LegacySegmentSpec;
