+++ processing/src/main/java/io/druid/query/search/RoaringBitmapDecisionHelper.java
@@ -17,7 +17,7 @@
  * under the License.
  */
 
-package io.druid.query.search.search;
+package io.druid.query.search;
 
 public class RoaringBitmapDecisionHelper extends SearchQueryDecisionHelper
 {
