+++ processing/src/main/java/io/druid/query/search/SearchQueryQueryToolChest.java
@@ -52,9 +52,6 @@ import io.druid.query.ResultMergeQueryRunner;
 import io.druid.query.aggregation.MetricManipulationFn;
 import io.druid.query.dimension.DimensionSpec;
 import io.druid.query.filter.DimFilter;
-import io.druid.query.search.search.SearchHit;
-import io.druid.query.search.search.SearchQuery;
-import io.druid.query.search.search.SearchQueryConfig;
 
 import javax.annotation.Nullable;
 import java.nio.ByteBuffer;
