+++ processing/src/main/java/io/druid/query/search/SearchQueryRunner.java
@@ -35,10 +35,6 @@ import io.druid.query.QueryRunner;
 import io.druid.query.Result;
 import io.druid.query.dimension.ColumnSelectorStrategy;
 import io.druid.query.dimension.ColumnSelectorStrategyFactory;
-import io.druid.query.search.search.SearchHit;
-import io.druid.query.search.search.SearchQuery;
-import io.druid.query.search.search.SearchQueryExecutor;
-import io.druid.query.search.search.SearchQuerySpec;
 import io.druid.segment.ColumnValueSelector;
 import io.druid.segment.DimensionSelector;
 import io.druid.segment.DoubleColumnSelector;
