+++ processing/src/test/java/io/druid/query/TestQueryRunners.java
@@ -26,7 +26,7 @@ import io.druid.collections.StupidPool;
 import io.druid.query.search.SearchQueryQueryToolChest;
 import io.druid.query.search.SearchQueryRunnerFactory;
 import io.druid.query.search.SearchStrategySelector;
-import io.druid.query.search.search.SearchQueryConfig;
+import io.druid.query.search.SearchQueryConfig;
 import io.druid.query.timeboundary.TimeBoundaryQueryRunnerFactory;
 import io.druid.query.timeseries.TimeseriesQueryEngine;
 import io.druid.query.timeseries.TimeseriesQueryQueryToolChest;
