+++ processing/src/test/java/io/druid/query/search/SearchQueryQueryToolChestTest.java
@@ -29,9 +29,6 @@ import io.druid.query.CacheStrategy;
 import io.druid.query.Druids;
 import io.druid.query.Result;
 import io.druid.query.TableDataSource;
-import io.druid.query.search.search.FragmentSearchQuerySpec;
-import io.druid.query.search.search.SearchHit;
-import io.druid.query.search.search.SearchQuery;
 import io.druid.query.spec.MultipleIntervalSegmentSpec;
 import org.junit.Assert;
 import org.junit.Test;
