+++ processing/src/test/java/io/druid/query/search/SearchQueryRunnerWithCaseTest.java
@@ -31,12 +31,6 @@ import io.druid.query.Druids;
 import io.druid.query.QueryPlus;
 import io.druid.query.QueryRunner;
 import io.druid.query.Result;
-import io.druid.query.search.search.AutoStrategy;
-import io.druid.query.search.search.CursorOnlyStrategy;
-import io.druid.query.search.search.SearchHit;
-import io.druid.query.search.search.SearchQuery;
-import io.druid.query.search.search.SearchQueryConfig;
-import io.druid.query.search.search.UseIndexesStrategy;
 import io.druid.segment.IncrementalIndexSegment;
 import io.druid.segment.QueryableIndex;
 import io.druid.segment.QueryableIndexSegment;
