+++ processing/src/test/java/io/druid/segment/SchemalessTestSimpleTest.java
@@ -42,8 +42,8 @@ import io.druid.query.aggregation.post.ArithmeticPostAggregator;
 import io.druid.query.aggregation.post.ConstantPostAggregator;
 import io.druid.query.aggregation.post.FieldAccessPostAggregator;
 import io.druid.query.search.SearchResultValue;
-import io.druid.query.search.search.SearchHit;
-import io.druid.query.search.search.SearchQuery;
+import io.druid.query.search.SearchHit;
+import io.druid.query.search.SearchQuery;
 import io.druid.query.spec.MultipleIntervalSegmentSpec;
 import io.druid.query.spec.QuerySegmentSpec;
 import io.druid.query.timeboundary.TimeBoundaryQuery;
