+++ processing/src/test/java/io/druid/segment/filter/SearchQueryFilterTest.java
@@ -34,8 +34,8 @@ import io.druid.js.JavaScriptConfig;
 import io.druid.query.extraction.ExtractionFn;
 import io.druid.query.extraction.JavaScriptExtractionFn;
 import io.druid.query.filter.SearchQueryDimFilter;
-import io.druid.query.search.search.ContainsSearchQuerySpec;
-import io.druid.query.search.search.SearchQuerySpec;
+import io.druid.query.search.ContainsSearchQuerySpec;
+import io.druid.query.search.SearchQuerySpec;
 import io.druid.segment.IndexBuilder;
 import io.druid.segment.StorageAdapter;
 import org.junit.AfterClass;
